requirements.txt
black
pytest
requests
python-decouple